modif post fait

+++ backend/controllers/post.js
@@ -88,14 +88,16 @@ exports.modifyPost = (req, res, next) => {
 
     postObject = { ...postObject, userId: userId, postId: id };
 
+    console.log(postObject.imageUrl)
+
     postModel.checkUserId(id, userId)
         .then(goodId => {
             postModel.findPhoto(id)
                 .then(oldPhoto => {
-                    if (oldPhoto[0].imageUrl !== null) {
-                        console.log('tut')
+                    if (oldPhoto[0].imageUrl !== null && postObject.imageUrl !== null) {
                         const filename = oldPhoto[0].imageUrl.split('/images/')[1];
                         fs.unlink(`images/${filename}`, () => {
+                            console.log('tatarar')
                             postModel.update(postObject)
                                 .then(result => {
                                     res.status(200).json({ result });
